backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
Pillow==10.1.0
reportlab==4.0.7
aiofiles==23.2.1
slowapi==0.1.9
redis==5.0.1
pytest==7.4.3
httpx==0.25.2
sentry-sdk[fastapi]==1.39.1
flake8==6.1.0
black==23.12.0
isort==5.13.2
mypy==1.7.1
bandit==1.7.5
safety==2.3.5
